Initial import of the galera reconciler
Let the mariadb operator reconcile galera resources, and update
the way it reconciles mariadbdatabases, based on whether mariadb
or galera is currently in use in the cluster.

It is assumed that either mariadb or galera is instantiated, you
can't run both at the same time and reconcile mariadbdatabases.

As the galera reconciler becomes more mature, we may want to
deprecate the mariadb reconciler and only support galera, as it
can be deployed as a 1-node statefulset.