Fix resource labels and selector in associated services
When creating multiple mariadb or galera CRs in the same environment
the associated services all have the same label selectors, which make
pods show up as endpoints on the wrong services.

Fix that by reworking the labels associated to pods/statefulsets,
and tweak the selectors in the service to only target the expected
resources.