galera: improve recovery of pods
The mariadb operator determines whether a starting pod will bootstrap
a galera cluster or join an existing one. On startup, the galera pod
receives that information via a temporary file injected by the operator.
Once galera is running in the pod, the operator updates its status.

Improve the status tracking so that the operator can detect when
a pod failed to start galera, and try again in a subsequent reconcile
loop.

Also make sure that the operator cannot inspect the galera state
while the galera server is being started, to prevent any corruption.