Use automatically generated namespace for kuttl
After recent changes in dynamic service account, since mariadb can be
deployed in any namespace, this change modifies the way the kuttl tests
are run, so they run in a namespace automatically created (and deleted) by
kuttl. This allows us to run the kuttl tests in a cluster where we have
already a deployed instance, without interfering. Additionally, we get
the benefit of being able to run multiple tests in parallel, since each
runs in its own namespace, which also reduces the runtime.