Simplify condition check in kuttl tests
Currently we assert every single field in conditions but this requires
updating the expected values in case any message is updated in
the underlying dependencies.

Because we have env test which is more suitable to assert conditions,
this simplifies the assertions in kuttl tests.

This also updates the tests for Galera CR to ensure the overall Ready
condition is updated successfully.