Use pre-created osp-secret in kuttl tests
Currently we create osp-secret in kuttl test but this conflicts with
the secret created by install_yamls. This removes the creation of
the secret in kuttl to avoid updating the existing secret.

This also updates the sample file to use that common secret, so that
we can use the file in kuttl tests without modification.