galera: support mariabackup SST method
Make the SST method configurable in the galera custom resource, and
allow both rsync and mariabackup methods.
Mariabackup requires a database user's credentials to operate, so
reuse the root db user for the time being.

The ability to inject passwords in template galera config files has
been implemented. We benefit from that to generate the usual
.my.cnf config file for mysql CLI running in the galera container.