when generating metadata.name, use dashes instead of underscores
The Database{} facade that is now used to generate a MariaDBDatabase
as well as a MariaDBAccount is currently generating a metadata
name for the MariaDBAccount based on the username.  As we do in
other places, if this username has underscores in it as is common,
swap them out with dashes.