dont err for missing Secret in GetDatabaseByNameAndAccount
The MariaDBAccount CR / controller currently does not add
finalizers to the Secret that's referenced by the controller.  This
creates a situation where the Secret can be deleted and fully
erased from the cluster, but the
MariaDBAccount still refers to that named secret.  When this happens, if a
controller is reconciling the delete of its own CR as well as its
MariaDBAccount, the lookup for MariaDBDatabase/MariaDBAccount must
succeed so that db.DeleteFinalizer can be called.   Prior to this
change, the GetDatabaseByNameAndAccount returns a not-found, causing
the calling controller to skip it and delete its own CR, leaving
the MariaDBAccount and MariaDBDatabase dangling.

the issue can be reproduced by building up an openstack env with
keystone / galera / memcached / rabbitmq, then deleting the namespace.
k8s will delete the secrets more quickly than it can remove the CRs
since they are reconciling the delete.