apply finalizers to Secret for MariaDBAccount
when a controller is reconciling the delete of its own CR as well as its
MariaDBAccount, the lookup for MariaDBDatabase/MariaDBAccount must
succeed so that db.DeleteFinalizer can be called.    This uses
GetDatabaseByNameAndAccount which is required to locate all three
of MariaDBDatabase, MariaDBAccount, and Secret, else notfound is
returned.

therefore, prevent the Secret from being prematurely removed from
the cluster by adding a calling CR finalizer to it, the same as
it's added to the MariaDBAccount.

Prior to this change, the GetDatabaseByNameAndAccount can return a not-found
even though the MariaDBAccount exists, causing
the calling controller to skip it and delete its own CR, leaving
the MariaDBAccount and MariaDBDatabase dangling.

the issue can be reproduced by building up an openstack env with
keystone / galera / memcached / rabbitmq, then deleting the namespace.
k8s will delete the secrets more quickly than it can remove the CRs
since they are reconciling the delete.