ensure galera service can be found before continuing
For whatever reason, services seem to run most of the time
barreling right into CreateOrPatchAll which immediately assumes
there is a Service object for Galera, which would provide for
a hostname. There does not appear to be any such
guarantee however, so CreateOrPatchAll hits an empty services list
(as would be expected if Galera is not reconciled yet), this would
just error out.

We can't wait for Galera to be bootstrapped here since none of the functional
test suites do any kind of SimulateGaleraBootstrapped kind of fixture,
and it seems like a bad idea to stuff all of that into
crd.CreateDBService.  So instead, a wait is set up strictly for the
service list fetch.