Add finalizer to galera ClusterIP service
When the service gets deleted manually and re-created it will
most likely receive a new ClusterIP. This results in services
trying to connect to the DB to get stuck.

To prevent manual deletion of the service this patch adds
a finalizer to the k8s service and removes it on galera service
delete.

Signed-off-by: Martin Schuppert <mschuppert@redhat.com>