Ability to configure a Galera CR to log on disk
Add the ability to store galera logs on disk, in addition
to logging to stdout.

When used, galera logs on disk, and an additional sidecar
container tail the output of the log file on stdout for
honouring regular kubectl logs command.

With this change, both the mysql database on disk and the logs
use the same PVC, and they each mount a subpath to distinguish
database and logs on disk. This way, storage configuration is
not impacted, the logs and the database are always collocated,
and logs are not stored under /var/log/mysql, so there is no
bad interaction with e.g. SST.

By default, logging is not enabled, so there is no operational
impact.

Jira: OSPRH-8209