Add CR name length validation
The galera controller creates StatefulSet for the db to run.
This adds a StatefulSet pod's label
"controller-revision-hash": "<statefulset_name>-<hash>" to the pod.
The kubernetes label is restricted under 63 char and the revision
hash is an int32, 10 chars + the hyphen. The galera controller also
adds a suffix '-galera' to the statefulset name. With this the max
name must not exceed 46 chars.

Depends-On: https://github.com/openstack-k8s-operators/lib-common/pull/532

Jira: https://issues.redhat.com/browse/OSPRH-8063

Signed-off-by: Martin Schuppert <mschuppert@redhat.com>