Fast bootstrap using safe_to_bootstrap flag
When a galera cluster is succesfully stopped, the last node to
shut down sets a internal flag called "safe_to_bootstrap", that
signals that this node should be the one to restart the cluster from.

This can help to improve bootstrap as once this flag is found on a
node, there is no need to inspect other nodes. This can also helps
to bootstrap a cluster in case e.g. various nodes are unavailable or
they suffered a db corruption. As soon as safe_to_bootstrap is read
on a node, the cluster can be restarted from it.

Change the way the mariadb operator extracts the sequence number
information from galera nodes, to retrieve additional flags such as
the "safe_to_bootstrap", and use it to improve the bootstrap node
selection.

Note that this requires that the cluster stops in a orchestrated way,
which is not always the case, for instance if all pods are deleted at
once. To get around it, add a PreStop hook script to sequentialize
shutdown of pods and their disconnection from the galera cluster.

Jira: OSPRH-10195