Do not use MustParse to process PVC size
The resource.MustParse function was previously used by the operator to
process the storageSize request provided by the top-level CR. However,
even though defer() and recover() can be used to handle the panic()
generated by that function in case of a wrong user input, that approach
doesn't seem to work properly and the operator crashes without any
chance to recover. The ParseQuantity implementation returns an error
instead of panic(), and it can be easily caught at the operator level.
This change moves away from the MustParse usage and relies on the
ParseQuantity mechanism.
".SpecCore" webhooks are updated to catch allErrs in case of a wrong
quantity string is detected, and [1] will push an update with the next
lib-common bump.

[1] https://github.com/openstack-k8s-operators/lib-common/pull/608

Signed-off-by: Francesco Pantano <fpantano@redhat.com>