New GaleraBackup CR to orchestrate backup of a running cluster
This CR sets up a cronjob whose purpose is to capture a consistent
snapshot of a running Galera node (via garbd) and to convert the
data internally into a regular SQL backup made with mysqldump.

The restore part is out of scope of this CR and will be addressed
in subsequent commits.

Jira: OSPRH-10503