fix harness to not look for secret finalizers
in af80f46c6b660789f774f2e4c40edbd822a1292d we removed the
client finalizer being added to the secret; it's added only
to MariaDBAccount, where the MariaDBAccount maintains its
own finalizer on the secret.  However, we forgot to update the
test harnesses in harnesses.go which check for client finalizers
in both MariaDBAccount as well as the Secret; this change updates
all those suites to not check inside the Secret since it no longer
receives the finalizer.